finalize version
